% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Generate Simulated Data}
\usage{
simData(n, p, mean = rep(0, p), Sigma = diag(p))
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of variables.}

\item{mean}{A vector of means for the variables (default is rep(0, p)).}

\item{Sigma}{The covariance matrix for the variables (default is diag(p)).}
}
\value{
A matrix of simulated data with dimensions n x p.
}
\description{
Generate simulated multivariate normal data.
}
\examples{
# Generate simulated data with default parameters
# Covariance structure with Compound Symmetry structure
n = 100
p = 3
rho = 0.3
cov_mat <- covMatCS(p = p, rho = rho)
sim_data <- simData(n = n, p = p, Sigma = cov_mat)
head(sim_data)
# Generate simulated data with custom mean and covariance matrix
custom_mean <- c(1, 2, 3)
sim_data_custom <- simData(n = 100, p = 3, mean = custom_mean, Sigma = cov_mat)
head(sim_data_custom)
}
