% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indTest.R
\name{indTest}
\alias{indTest}
\title{Independent Test for Covariance Matrix}
\usage{
indTest(X, covMat = NULL, alpha = 0.05)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the variables.}

\item{covMat}{Optional. A numeric matrix representing the covariance matrix to be tested against. If NULL, the sample covariance matrix is used (default is NULL).}

\item{alpha}{The significance level for the test (default is 0.05).}
}
\value{
A data frame containing the test statistic, degrees of freedom, alpha value, p-value, and test result.
#' @references
Marques, F. J., Diogo, J., Norouzirad, M., & Bispo, R. (2023). Testing the independence of variables for specific covariance structures: A simulation study. Mathematical Methods in the Applied Sciences, 46(9), 10421–10434. DOI: 10.1002/mma.9130
}
\description{
Performs an independent test for the covariance matrix to assess if it is significantly different from a given covariance matrix or the sample covariance matrix.
}
\examples{
# Example usage:

library(MASS)

n = 100 # Sample Size
p = 5
rho = 0.014
# Covariance structure with Autoregressive structure
cov_mat <- covMatAR(p = p, rho = rho)
data <- mvrnorm(n = n, mu = rep(0,p), Sigma = cov_mat)
indTest(data)
indTest(data, covMat = cov_mat)

# Covariance structure with Compound Symmetry structure
cov_mat <- covMatCS(p = p, rho = rho)
data <- mvrnorm(n = n, mu = rep(0,p), Sigma = cov_mat)
indTest(data)
indTest(data, covMat = cov_mat)

#' # Covariance structure with Circular structure
cov_mat <- covMatC(p = p, rho = rho)
data <- mvrnorm(n = n, mu = rep(0,p), Sigma = cov_mat)
indTest(data)
indTest(data, covMat = cov_mat)


}
